Another round of code review changes. Includes:
- Adding UPPERCASE aliases to HDF5 records
- Fix the type of all records for int/float distinction
- Many TODO notes for more work
- Fix up GetChanges in_error handling
- Resolved many TODOs (most were no-ops)

+++ pandablocks/responses.py
@@ -75,7 +75,7 @@ class ScalarFieldInfo(FieldInfo):
 
     units: Optional[str] = None
     scale: Optional[float] = None
-    offset: Optional[int] = None  # TODO: PandA returns 0 - is this an int or a float 0?
+    offset: Optional[float] = None
 
 
 @dataclass
